+++ apps/client-web/pages/AppLayout/AppLayout.tsx
@@ -4,9 +4,9 @@ import {Button, Layout, Space} from 'antd';
 import {Link, Route, Switch} from 'react-router-dom';
 
 import {AppContentSwitch} from '../AppContentSwitch/AppContentSwitch';
-import AppRouteCollection from '../../../constants/AppRouteCollection';
-import {LandingScreen} from '../../LandingScreen/LandingScreen';
-import {LocationBreadcrumb} from '../../LocationBreadcrumb/LocationBreadcrumb';
+import AppRouteCollection from '../../constants/AppRouteCollection';
+import {LandingScreen} from '../../components/LandingScreen/LandingScreen';
+import {LocationBreadcrumb} from '../../components/LocationBreadcrumb/LocationBreadcrumb';
 import React from 'react';
 
 const {Header, Content, Footer} = Layout;
