--- apps/client-web/pages/authentication-demo.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-
-/** The home or landing page of the application. */
-export function Index() {
-  return <></>;
-}
-
-export default Index;
