--- apps/client-web/pages/index.module.scss
@@ -1,5 +1,0 @@
-.pageBackground {
-  background-image: url('/waves.svg');
-  background-size: 100% auto;
-  background-repeat: no-repeat;
-}
