--- apps/client-web/styles/main.less
@@ -1,19 +1,0 @@
-// a common class to maximise a node's height
-.max-height {
-  height: 100%;
-}
-
-// an class for banner images, etc.
-.image-cover {
-  width: 100%;
-  height: 100%;
-  background-repeat: no-repeat;
-  background-position: center center;
-  background-size: cover;
-  object-fit: cover;
-}
-
-.image-max {
-  max-width: 100%;
-  max-height: 100%;
-}
